Move pages dir into components dir.
I did this for two main reasons:

* I suspect that the distinction between 'pages' and 'components' will
  become blurry as the site evolves and becomes more dynamic. In fact,
  in some sense the distinction already *is* blurry. If things do
  become too complicated in the future, though, we can always go back
  to using the 'pages' directory.

* Pages use lots of components, and having to start every component's
  `require()` with `../components/` was getting quite tedious.
